chore(ui): move Vite UI version define under scripts/
Place getUiImportMetaEnvDefine next to other build scripts so ESLint's
existing scripts/** TypeScript block applies without special-casing repo-root
config files. Vite and Vitest import ./scripts/vite-ui-env-define.js.
