+++ vitest.config.ts
@@ -1,7 +1,7 @@
 import path from "node:path";
 import { defineConfig } from "vitest/config";
 import react from "@vitejs/plugin-react";
-import { getUiImportMetaEnvDefine } from "./vite-ui-env-define.js";
+import { getUiImportMetaEnvDefine } from "./scripts/vite-ui-env-define.js";
 
 const vitestReporters = process.env.CI ? ["default", "github-actions"] : ["default"];
